+++ client/src/App.tsx
@@ -36,7 +36,6 @@ function App() {
   const dispatch = useDispatch();
   const checkLogin = async () => {
     const token = await getToken();
-    console.log('token: ', token);
     if (!token) return;
     const response = await API.post('/auth/autologin', '', {
       headers: {
