+++ client/src/components/Meal/MenuBar/index.tsx
@@ -1,17 +1,30 @@
 import React from 'react';
+import { getImageURL } from '@utils/util';
 import { Button, Message } from 'semantic-ui-react';
 import { ResizeHeaderProps } from '@set/ResizeHeader';
-import { getImageURL } from '@utils/util';
 import QrCode from '@ingredients/QrCode';
-import useMenuBar from './useMenuBar';
 import MenuBarInput from '@meal/MenuBarInput';
 import MenuBarOutput from '@meal/MenuBarOutput';
 import MenuBarButtons from '@meal/MenuBarButtons';
+import useSaveButtons from '@set/SaveButtons/useSaveButtons';
 import * as S from './style';
 
 function MenuBar({ onToggle }: ResizeHeaderProps) {
-  const { message, Modal, imageUrl } = useMenuBar();
-
+  const {
+    message,
+    Modal,
+    imageUrl,
+    downloadImage,
+    downloadText,
+    createHandler,
+    clipboardHandler,
+  } = useSaveButtons();
+  const MenuBarSaveButtons = [
+    { title: '이미지 저장', iconName: 'image', size: 'massive', onClickHandler: downloadImage },
+    { title: '텍스트 저장', iconName: 'file text', size: 'massive', onClickHandler: downloadText },
+    { title: 'QR코드 생성', iconName: 'qrcode', size: 'massive', onClickHandler: createHandler },
+    { title: '화면 적용', iconName: 'tv', size: 'massive', onClickHandler: clipboardHandler },
+  ];
   return (
     <S.MenuBarContainer>
       <S.Logo>
@@ -19,7 +32,7 @@ function MenuBar({ onToggle }: ResizeHeaderProps) {
       </S.Logo>
       <MenuBarInput />
       <MenuBarOutput />
-      <MenuBarButtons />
+      <MenuBarButtons MenuBarSaveButtons={MenuBarSaveButtons} />
       <S.ButtonWrapper>
         <Button.Group basic size="massive">
           <Button icon="sort" size="massive" onClick={onToggle} />
