+++ client/src/components/Meal/MenuBar/style.ts
@@ -26,7 +26,6 @@ export const Logo = styled.div`
   height: 100%;
   width: 65px;
   margin-left: 20px;
-  margin-right: 10px;
 `;
 
 export const LogoImg = styled.img`
