+++ client/src/components/Set/SaveButtons/useSaveButtons.ts
@@ -4,13 +4,12 @@ import { RootState } from '@contexts/index';
 import html2canvas from 'html2canvas';
 import * as clipboard from 'clipboard-polyfill';
 import { ClipboardItem } from 'clipboard-polyfill';
-import { useEffect, useRef, useState } from 'react';
-import { getToken, setToken } from '@utils/token';
+import { useRef, useState } from 'react';
+import { setToken } from '@utils/token';
 import useToggle from '@hooks/useToggle';
 import { BASE_URL } from '@lib/apis/common';
 import useModal from '@hooks/useModal';
 import { userLogin, userLogout } from '@contexts/user';
-import { API } from '@lib/apis/common';
 import axios from 'axios';
 
 export const useSaveButtons = () => {
